Remove dependence on the Rust source tree and stop building libcore.
To generate sufficiently small code, we have to build with LTO, and with
LTO, all the Rust code is combined first before LLVM compiles it, including
libcore.  Therefore, there's no need to compile libcore separately with
size optimizations.  By removing it from the build, we can build this
project using only an i686 Rust distribution (as long as it has
-C opt-level=s).